tests/wpt/tests/html/canvas/offscreen/canvas-host/2d.canvas.host.initial.color.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.canvas.host.initial.color</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.host.initial.color</h1>
<p class="desc">Initial state is transparent black</p>


<script>
var t = async_test("Initial state is transparent black");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  _assertPixel(canvas, 20,20, 0,0,0,0);
  t.done();

});
</script>
